Inicio da correcao de buscar por Feign ensinada pelo Tales
Inicio da correcao de buscar por Feign ensinada pelo Tales

+++ src/main/java/br/edu/ulbra/election/party/service/PartyService.java
@@ -6,6 +6,9 @@ import br.edu.ulbra.election.party.input.v1.PartyInput;
 import br.edu.ulbra.election.party.model.Candidate;
 import br.edu.ulbra.election.party.model.Party;
 import br.edu.ulbra.election.party.output.v1.CandidateOutput;
+import br.edu.ulbra.election.party.output.v1.ElectionOutput;
+import br.edu.ulbra.election.party.client.ElectionClientService;
+import br.edu.ulbra.election.party.client.PartyClientService;
 import br.edu.ulbra.election.party.output.v1.GenericOutput;
 import br.edu.ulbra.election.party.output.v1.PartyOutput;
 import br.edu.ulbra.election.party.repository.CandidateRepository;
@@ -17,6 +20,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,13 +32,17 @@ public class PartyService {
     private final PartyRepository partyRepository;
     private final CandidateClientService candidateClientService;
     private final CandidateRepository candidateRepository;
+    private final ElectionClientService electionClientService;
+    private final PartyClientService partyClientService;
     private final ModelMapper modelMapper;
 
     @Autowired
-    public PartyService(PartyRepository partyRepository, CandidateClientService candidateClientService, CandidateRepository candidateRepository, ModelMapper modelMapper){
+    public PartyService(PartyRepository partyRepository, CandidateClientService candidateClientService, CandidateRepository candidateRepository, PartyClientService partyClientService, ElectionClientService electionClientService, ModelMapper modelMapper){
         this.partyRepository = partyRepository;
         this.candidateClientService = candidateClientService;
         this.candidateRepository = candidateRepository;
+        this.electionClientService = electionClientService;
+        this.partyClientService = partyClientService;
         this.modelMapper = modelMapper;
     }
 
@@ -42,6 +50,17 @@ public class PartyService {
         List<Party> partyList = (List<Party>)partyRepository.findAll();
         return partyList.stream().map(Party::toPartyOutput).collect(Collectors.toList());
     }
+
+
+    public List<CandidateOutput> getAllByPartyId(){
+        try {
+            List<Candidate> candidateList = (List<Candidate>) candidateRepository.findAll();
+            return candidateList.stream().map(this::toCandidateOutput).collect(Collectors.toList());
+        } catch (Exception e) {
+            throw new GenericOutputException(MESSAGE_INVALID_TO_DELETE_PARTY);
+        }
+    }
+
 
     public PartyOutput create(PartyInput partyInput) {
         validateInput(partyInput);
@@ -88,8 +107,9 @@ public class PartyService {
             throw new GenericOutputException(MESSAGE_INVALID_ID);
         }
 
-        Party party = partyRepository.findById(partyId).orElse(null);
-        if (party == null){
+//        Party party = partyRepository.findById(partyId).orElse(null);
+        Party party = partyRepository.findFirstById(partyId);
+        if (party.getId() == null){
             throw new GenericOutputException(MESSAGE_PARTY_NOT_FOUND);
         }
 
@@ -97,7 +117,7 @@ public class PartyService {
 
         /* Não pode ser excluído um partido com candidatos */
 
-            CandidateOutput candidateOutput;
+/*            CandidateOutput candidateOutput;
             for(int i = 1; i<1000; i++) {
 
         try {
@@ -115,6 +135,40 @@ public class PartyService {
         }
 
             }
+
+*/
+
+
+
+        try{
+            CandidateOutput candidateOutput = candidateClientService.getPartyId(partyId);
+
+
+            if(candidateOutput.getPartyId() == partyId) {
+                System.out.println(candidateOutput.getPartyId());
+            } /* Teste */
+            throw new GenericOutputException("Erro proposital!");
+
+        } catch (FeignException e){
+            if (e.status() == 500) {
+                throw new GenericOutputException("Invalid Party");
+            }
+        }
+
+
+
+
+
+
+/*
+        if (candidateOutput.getPartyId().toString().equals(party.getId().toString())){
+            throw new GenericOutputException(MESSAGE_INVALID_TO_DELETE_PARTY);
+        }
+*/
+
+
+//Fazer por getByPartyId;
+
 /*********************/
 
         partyRepository.delete(party);
@@ -144,5 +198,17 @@ public class PartyService {
             throw new GenericOutputException("Invalid Party Number");
         }
     }
+
+
+    public CandidateOutput toCandidateOutput(Candidate candidate){
+        CandidateOutput candidateOutput = modelMapper.map(candidate, CandidateOutput.class);
+        ElectionOutput electionOutput = electionClientService.getById(candidate.getElectionId());
+        candidateOutput.setElectionOutput(electionOutput);
+        PartyOutput partyOutput = partyClientService.getById(candidate.getPartyId());
+        candidateOutput.setPartyOutput(partyOutput);
+        return candidateOutput;
+    } /* Adição*/
+
+
 
 }
